+++ app/Http/Controllers/TaxonomyController.php
@@ -11,14 +11,13 @@ class TaxonomyController extends Controller
 {
     /**
      * All Utils instance.
-     *
      */
     protected $moduleUtil;
 
     /**
      * Constructor
      *
-     * @param ProductUtils $product
+     * @param  ProductUtils  $product
      * @return void
      */
     public function __construct(ModuleUtil $moduleUtil)
@@ -34,39 +33,73 @@ class TaxonomyController extends Controller
     public function index()
     {
         $category_type = request()->get('type');
-        if ($category_type == 'product' && !auth()->user()->can('category.view') && !auth()->user()->can('category.create')) {
+        if ($category_type == 'product' && ! auth()->user()->can('category.view') && ! auth()->user()->can('category.create')) {
             abort(403, 'Unauthorized action.');
         }
 
         if (request()->ajax()) {
+            $can_edit = true;
+            if ($category_type == 'product' && ! auth()->user()->can('category.update')) {
+                $can_edit = false;
+            }
+
+            $can_delete = true;
+            if ($category_type == 'product' && ! auth()->user()->can('category.delete')) {
+                $can_delete = false;
+            }
+
             $business_id = request()->session()->get('user.business_id');
 
-            $category = Category::where('business_id', $business_id)
-                            ->where('category_type', $category_type)
-                            ->select(['name', 'short_code', 'description', 'id', 'parent_id']);
+            $all_categories = Category::where('business_id', $business_id)
+                ->where('category_type', $category_type)
+                ->get()
+                ->keyBy('id');
+
+            $grouped = $all_categories->groupBy('parent_id');
+
+            $category = collect();
+
+            // Get parents (those with parent_id = null or 0)
+            $parents = $grouped[null] ?? $grouped[0] ?? [];
+
+            foreach ($parents as $parent) {
+                // Push parent
+                $category->push($parent);
+
+                // Get and push children with prefixed parent name
+                foreach ($grouped[$parent->id] ?? [] as $child) {
+                    // Attach parent name for use in editColumn
+                    $child->parent_name = $parent->name;
+                    $category->push($child);
+                }
+            }
 
             return Datatables::of($category)
-                ->addColumn(
-                    'action',
-                    '
-                    <button data-href="{{action(\'TaxonomyController@edit\', [$id])}}?type=' . $category_type . '" class="btn btn-xs btn-primary edit_category_button"><i class="glyphicon glyphicon-edit"></i>  @lang("messages.edit")</button>
-                        &nbsp;
-                    
-                        <button data-href="{{action(\'TaxonomyController@destroy\', [$id])}}" class="btn btn-xs btn-danger delete_category_button"><i class="glyphicon glyphicon-trash"></i> @lang("messages.delete")</button>
-                    '
-                )
-                ->editColumn('name', function ($row) {
-                    if ($row->parent_id != 0) {
-                        return '--' . $row->name;
-                    } else {
-                        return $row->name;
+                ->addColumn('action', function ($row) use ($can_edit, $can_delete, $category_type) {
+                    $html = '';
+                    if ($can_edit) {
+                        $html .= '<button data-href="' . action([\App\Http\Controllers\TaxonomyController::class, 'edit'], [$row->id]) . '?type=' . $category_type . '" class="tw-dw-btn tw-dw-btn-xs tw-dw-btn-outline tw-dw-btn-primary edit_category_button"><i class="glyphicon glyphicon-edit"></i> ' . __('messages.edit') . '</button>';
                     }
+
+                    if ($can_delete) {
+                        $html .= '&nbsp;<button data-href="' . action([\App\Http\Controllers\TaxonomyController::class, 'destroy'], [$row->id]) . '" class="tw-dw-btn tw-dw-btn-outline tw-dw-btn-xs tw-dw-btn-error delete_category_button"><i class="glyphicon glyphicon-trash"></i> ' . __('messages.delete') . '</button>';
+                    }
+
+                    return $html;
+                })
+                ->editColumn('name', function ($row) {
+                    // If parent_name is set (means it's a child)
+                    if (!empty($row->parent_name)) {
+                        return $row->parent_name . ' -> ' . $row->name;
+                    }
+                    return $row->name;
                 })
                 ->removeColumn('id')
                 ->removeColumn('parent_id')
                 ->rawColumns(['action'])
                 ->make(true);
-        }
+
+            }
 
         $module_category_data = $this->moduleUtil->getTaxonomyData($category_type);
 
@@ -81,7 +114,7 @@ class TaxonomyController extends Controller
     public function create()
     {
         $category_type = request()->get('type');
-        if ($category_type == 'product' && !auth()->user()->can('category.view') && !auth()->user()->can('category.create')) {
+        if ($category_type == 'product' && ! auth()->user()->can('category.create')) {
             abort(403, 'Unauthorized action.');
         }
         $business_id = request()->session()->get('user.business_id');
@@ -95,7 +128,7 @@ class TaxonomyController extends Controller
                         ->get();
 
         $parent_categories = [];
-        if (!empty($categories)) {
+        if (! empty($categories)) {
             foreach ($categories as $category) {
                 $parent_categories[$category->id] = $category->name;
             }
@@ -114,13 +147,13 @@ class TaxonomyController extends Controller
     public function store(Request $request)
     {
         $category_type = request()->input('category_type');
-        if ($category_type == 'product' && !auth()->user()->can('category.view') && !auth()->user()->can('category.create')) {
+        if ($category_type == 'product' && ! auth()->user()->can('category.create')) {
             abort(403, 'Unauthorized action.');
         }
 
         try {
             $input = $request->only(['name', 'short_code', 'category_type', 'description']);
-            if (!empty($request->input('add_as_sub_cat')) &&  $request->input('add_as_sub_cat') == 1 && !empty($request->input('parent_id'))) {
+            if (! empty($request->input('add_as_sub_cat')) && $request->input('add_as_sub_cat') == 1 && ! empty($request->input('parent_id'))) {
                 $input['parent_id'] = $request->input('parent_id');
             } else {
                 $input['parent_id'] = 0;
@@ -130,15 +163,15 @@ class TaxonomyController extends Controller
 
             $category = Category::create($input);
             $output = ['success' => true,
-                            'data' => $category,
-                            'msg' => __("category.added_success")
-                        ];
+                'data' => $category,
+                'msg' => __('category.added_success'),
+            ];
         } catch (\Exception $e) {
-            \Log::emergency("File:" . $e->getFile(). "Line:" . $e->getLine(). "Message:" . $e->getMessage());
-            
+            \Log::emergency('File:'.$e->getFile().'Line:'.$e->getLine().'Message:'.$e->getMessage());
+
             $output = ['success' => false,
-                            'msg' => __("messages.something_went_wrong")
-                        ];
+                'msg' => __('messages.something_went_wrong'),
+            ];
         }
 
         return $output;
@@ -164,14 +197,14 @@ class TaxonomyController extends Controller
     public function edit($id)
     {
         $category_type = request()->get('type');
-        if ($category_type == 'product' && !auth()->user()->can('category.view') && !auth()->user()->can('category.create')) {
+        if ($category_type == 'product' && ! auth()->user()->can('category.update')) {
             abort(403, 'Unauthorized action.');
         }
 
         if (request()->ajax()) {
             $business_id = request()->session()->get('user.business_id');
             $category = Category::where('business_id', $business_id)->find($id);
-            
+
             $module_category_data = $this->moduleUtil->getTaxonomyData($category_type);
 
             $parent_categories = Category::where('business_id', $business_id)
@@ -180,12 +213,12 @@ class TaxonomyController extends Controller
                                         ->where('id', '!=', $id)
                                         ->pluck('name', 'id');
             $is_parent = false;
-            
+
             if ($category->parent_id == 0) {
                 $is_parent = true;
                 $selected_parent = null;
             } else {
-                $selected_parent = $category->parent_id ;
+                $selected_parent = $category->parent_id;
             }
 
             return view('taxonomy.edit')
@@ -197,26 +230,27 @@ class TaxonomyController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int $id
+     * @param  int  $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
-
         if (request()->ajax()) {
             try {
                 $input = $request->only(['name', 'description']);
                 $business_id = $request->session()->get('user.business_id');
 
                 $category = Category::where('business_id', $business_id)->findOrFail($id);
+
+                if ($category->category_type == 'product' && ! auth()->user()->can('category.update')) {
+                    abort(403, 'Unauthorized action.');
+                }
+
                 $category->name = $input['name'];
                 $category->description = $input['description'];
+                $category->short_code = $request->input('short_code');
 
-                if (!empty($request->input('short_code'))) {
-                    $category->short_code = $request->input('short_code');
-                }
-                
-                if (!empty($request->input('add_as_sub_cat')) &&  $request->input('add_as_sub_cat') == 1 && !empty($request->input('parent_id'))) {
+                if (! empty($request->input('add_as_sub_cat')) && $request->input('add_as_sub_cat') == 1 && ! empty($request->input('parent_id'))) {
                     $category->parent_id = $request->input('parent_id');
                 } else {
                     $category->parent_id = 0;
@@ -224,14 +258,14 @@ class TaxonomyController extends Controller
                 $category->save();
 
                 $output = ['success' => true,
-                            'msg' => __("category.updated_success")
-                            ];
+                    'msg' => __('category.updated_success'),
+                ];
             } catch (\Exception $e) {
-                \Log::emergency("File:" . $e->getFile(). "Line:" . $e->getLine(). "Message:" . $e->getMessage());
-            
+                \Log::emergency('File:'.$e->getFile().'Line:'.$e->getLine().'Message:'.$e->getMessage());
+
                 $output = ['success' => false,
-                            'msg' => __("messages.something_went_wrong")
-                        ];
+                    'msg' => __('messages.something_went_wrong'),
+                ];
             }
 
             return $output;
@@ -246,23 +280,27 @@ class TaxonomyController extends Controller
      */
     public function destroy($id)
     {
-
         if (request()->ajax()) {
             try {
                 $business_id = request()->session()->get('user.business_id');
 
                 $category = Category::where('business_id', $business_id)->findOrFail($id);
+
+                if ($category->category_type == 'product' && ! auth()->user()->can('category.delete')) {
+                    abort(403, 'Unauthorized action.');
+                }
+
                 $category->delete();
 
                 $output = ['success' => true,
-                            'msg' => __("category.deleted_success")
-                            ];
+                    'msg' => __('category.deleted_success'),
+                ];
             } catch (\Exception $e) {
-                \Log::emergency("File:" . $e->getFile(). "Line:" . $e->getLine(). "Message:" . $e->getMessage());
-            
+                \Log::emergency('File:'.$e->getFile().'Line:'.$e->getLine().'Message:'.$e->getMessage());
+
                 $output = ['success' => false,
-                            'msg' => __("messages.something_went_wrong")
-                        ];
+                    'msg' => __('messages.something_went_wrong'),
+                ];
             }
 
             return $output;
@@ -275,11 +313,11 @@ class TaxonomyController extends Controller
             $api_token = request()->header('API-TOKEN');
 
             $api_settings = $this->moduleUtil->getApiSettings($api_token);
-            
+
             $categories = Category::catAndSubCategories($api_settings->business_id);
         } catch (\Exception $e) {
-            \Log::emergency("File:" . $e->getFile(). "Line:" . $e->getLine(). "Message:" . $e->getMessage());
-            
+            \Log::emergency('File:'.$e->getFile().'Line:'.$e->getLine().'Message:'.$e->getMessage());
+
             return $this->respondWentWrong($e);
         }
 
@@ -289,6 +327,7 @@ class TaxonomyController extends Controller
     /**
      * get taxonomy index page
      * through ajax
+     *
      * @return \Illuminate\Http\Response
      */
     public function getTaxonomyIndexPage(Request $request)
